+++ openmp/libomptarget/deviceRTLs/interface.h
@@ -462,9 +462,9 @@ EXTERN void __kmpc_get_team_static_memory(int16_t isSPMDExecutionMode,
 EXTERN void __kmpc_restore_team_static_memory(int16_t isSPMDExecutionMode,
                                               int16_t is_shared);
 
-// Device stack trace functions TODO modify datatype here to match needs
-EXTERN void omp_stack_trace_push(int32_t data);
-EXTERN int omp_stack_trace_pop(int32_t * data);
+// Device stack trace functions
+EXTERN void omp_stack_trace_push(int64_t data);
+EXTERN int omp_stack_trace_pop(int64_t * data);
 //EXTERN void omp_stack_trace_push(uint64_t Index);
 //EXTERN void omp_stack_trace_pop();
